add BadgeNotif to SelectCohort
