+++ packages/ds/src/admin/index.ts
@@ -4,7 +4,7 @@
 import InputText from "./form/InputText";
 import InputPhone from "./form/InputPhone";
 import Label from "./form/Label";
-import Select from "./form/Select";
+import Select from "./form/Select/Select";
 
 // Layout
 import Container from "./layout/Container";
@@ -15,6 +15,7 @@ import Navbar from "./layout/Navbar";
 
 // UI
 import Badge from "./ui/Badge";
+import BadgeNotif from "./ui/BadgeNotif";
 import Button from "./ui/Button";
 import DropdownButton from "./ui/DropdownButton";
 import Modal from "./ui/Modal";
@@ -35,6 +36,7 @@ export {
 
   // UI
   Badge,
+  BadgeNotif,
   Button,
   DropdownButton,
   Modal,
